change paged query total number of elements count to long and add constructor to simplify CernPage initialization

+++ src/main/java/ch/cern/todo/model/api/PageApi.java
@@ -32,5 +32,5 @@ public class PageApi<T extends Serializable> implements Serializable {
     /**
      * {@link PageImpl#getTotalPages()}
      */
-    private int totalPages;
+    private long totalPages;
 }
