+++ src/main/java/ch/cern/todo/repository/TaskRepositoryCustom.java
@@ -7,7 +7,7 @@ import ch.cern.todo.model.database.TaskEntity;
 
 public interface TaskRepositoryCustom {
 
-    int countTasks(SearchTask searchTask, LoggedUserInfo loggedUserInfo);
+    long countTasks(SearchTask searchTask, LoggedUserInfo loggedUserInfo);
 
     CernPage<TaskEntity> findTasks(SearchTask searchTask,  LoggedUserInfo loggedUserInfo);
 }
